xml/parsing/manakai/charref-3.dat

#data
<?xml version="1.0"?><p>&#x40"</p>
#errors
1;30;no refc;;;m
#document
| <{} p>
|   "@""
